+++ src/main/java/io/github/dataforseo/client/model/OnPageHtmlResourceItem.java
@@ -87,9 +87,9 @@ public class OnPageHtmlResourceItem  extends BaseOnPageResourceItem  {
 
   public static final String SERIALIZED_NAME_ONPAGE_SCORE = "onpage_score";
   @SerializedName(SERIALIZED_NAME_ONPAGE_SCORE)
-  private Double onpageScore;
+  private Integer onpageScore;
 
-  public OnPageHtmlResourceItem onpageScore(Double onpageScore) {
+  public OnPageHtmlResourceItem onpageScore(Integer onpageScore) {
     this.onpageScore = onpageScore;
     return this;
   }
@@ -102,11 +102,11 @@ public class OnPageHtmlResourceItem  extends BaseOnPageResourceItem  {
    * @return onpageScore
    */
   @javax.annotation.Nullable
-  public Double getOnpageScore() {
+  public Integer getOnpageScore() {
     return onpageScore;
   }
 
-  public void setOnpageScore(Double onpageScore) {
+  public void setOnpageScore(Integer onpageScore) {
     this.onpageScore = onpageScore;
   }
